wpr: Increase wpr server wait timeout to 2 minutes
This is a speculative fix for WPR wait timeouts seen on a number of
platforms. WPR in replay mode loads and parses an archive on startup
before starting its servers. This load process is very CPU intensive,
decompressing the archive into memory and parsing the JSON. As of
writing, the archive is 175M which also adds IO pressure to the startup.

On Grunt, with no load, this entire process at least 10 seconds. With
additional CPU and IO load (or throttling), it could take much longer.

BUG=b:195854568
TEST=tas run $DUT platform.MemoryPressureModerate

Change-Id: I4294abaa8ccd15fb182efcb7486d23d3081e5604