fingerprint: Fix InitializeKnownState
This changes the assumption that the rollback block id should
always be 1 (freshly flashed and entropy initialized).
This assumption is not how the rest of the fingerprint system operates.

This change further allows this function to actually repair the FPMCU
FW state from most of the problem scenarios, instead of immediately
erroring out.

BUG=b:195989768, b:194500081
TEST=../tast/fast_build.sh -T
TEST=tast run localhost:2222 firmware.Fp{TpmSeed,AddEntropy,\
CheckWriteProtect,SystemIsLocked,SoftwareWriteProtect,RDP0,\
RebootToRO,ReadFlash}
     # Ensure all pass

Change-Id: I78416afdc2c60c9afd92cd3ce52ac360668a10a9