fingerprint: Fix InitializeKnownState
There are a few faulty assumptions that InitializeKnownState
function makes. The most detrimental one is that the rollback
block ID should always be 1, even when it doesn't flash the fpmcu.

This changes the assumption that the rollback block id should
always be 1 (freshly flashed and entropy initialized).
This assumption is not how the rest of the fingerprint system operates.

This change further allows this function to actually repair the FPMCU
FW state from most of the problem scenarios, instead of immediately
erroring out.

BUG=b:195989768, b:194500081, b:196182618
TEST=../tast/fast_build.sh -T
TEST=ssh localhost -p2222 -- flash_fp_mcu
     sleep 5 # biod restarting will init entropy
     ssh localhost -p2222 -- ectool --name=cros_fp rollbackinfo
     # Block ID should be 1
     ssh localhost -p2222 -- bio_wash; sleep 1
     ssh localhost -p2222 -- ectool --name=cros_fp rollbackinfo
     # Block ID should be 3
     #
     # Comment out the "ReimageFPMCU" call in FirmwareTest.Close,
     # since this is not needed and defeats our work.
     #
     # Since the flash protect flags checking is broken, we
     # need to schedule in terms of keeping sw-wp
     time tast run localhost:2222 firmware.FpRDP0
     time tast run localhost:2222 firmware.Fp{AddEntropy,\
SystemIsLocked,SoftwareWriteProtect,RebootToRO,ReadFlash}
     # Ensure all pass

Change-Id: I78416afdc2c60c9afd92cd3ce52ac360668a10a9