fingerprint: Fix InitializeKnownState
This changes the assumption that the rollback block id should
always be 1 (freshly flashed and entropy initialized).
This assumption is not how the rest of the fingerprint system operates.

This change further allows this function to actually repair the FPMCU
FW state from most of the problem scenarios, instead of immediately
erroring out.

BUG=b:195989768, b:194500081
TEST=../tast/fast_build.sh -T
TEST=ssh localhost -p2222 -- flash_fp_mcu
     ssh localhost -p2222 -- bio_wash --factory_init; sleep 1
     ssh localhost -p2222 -- ectool --name=cros_fp rollbackinfo
     # Block ID should be 1
     ssh localhost -p2222 -- bio_wash; sleep 1
     ssh localhost -p2222 -- ectool --name=cros_fp rollbackinfo
     # Block ID should be 3
     time tast run localhost:2222 firmware.Fp{TpmSeed,AddEntropy,\
CheckWriteProtect,SystemIsLocked,SoftwareWriteProtect,RDP0,\
RebootToRO,ReadFlash}
     # Ensure all pass

Change-Id: I78416afdc2c60c9afd92cd3ce52ac360668a10a9