tast-tests: Rename /var/lib/whitelist
Make tests use /var/lib/devicesettings instead. Currently, that's a bind
mount to /var/lib/whitelist. Once the migration is finished it will be
the normal directory to use.
We need to be careful when removing the directory now. If we just delete
/var/lib/whitelist, the bind mount will break and it would only be
created on reboot. That's why we now delete the content but not the
folder.

BUG=b:187793661
TEST=Run the following tast tests: OwnershipAPI, OwnershipNotRetaken,
     SELinuxFilesSystem, NoteEnterpriseOwned, OwnershipRetaken,
     ReadErrorLog, KernelCrash (with --extrauseflags chrome_internal)

Cq-Depend: 3088869
Change-Id: I135a4715d0daf4bc83ef02599d7bc244316a4f1a