Add an option to specify the encrypted real-time event reporting server.
In preparation for the integration test for the Chrome/ChromeOS
enterprise encrypted real-time event reporting server, this change adds an option
in the Tast test to start Chrome and point Chrome to report real time
events to a specified encrypted real time server address.

The integration test will use this option to direct Chrome to report
events to the alpha encrypted real time event reporting server.

BUG=b:187795673
TEST=None

Change-Id: Icb002c0445d72399cec5ca5785f493c5a9cc64ed