Add a new fixture to clear enrollment state on a device.
To support the real-time event reporting integration test, and other
tests that requires the DUT to enroll with real Device Management
servers, this change adds a fixture that wipes the enrollment state from
a device. A test device that enterprise enrolls can reset its state
before and after the test, and be restored to its original starting
state.

BUG=b:187795673
TEST=tested on a prototype real-time event reporting integration test.

Change-Id: If821303eec12bcf5a4eac89035818b70d3dca304