tast-tests: Ensure display is on before starting SplitView test
This CL ensures the display is on before starting the SplitView.
The display needs to be active in the test because the touch display input doesn’t work properly when the display turns off, but the display sometimes turns off due to the timeout while running a chain of tests with the same precondition.

BUG=b:196007379
TEST=tast run arc.SplitView.*

Change-Id: I1dea6fbf1a04aaf20816675a8c183325458ee5a7