fwupd: Expanding firmware.FwupdPowerdStartup to work with all plugins
Changes the command run by this test so that the output comes in json
format. This output is then parsed into a struct, which is iterated over
until the target plugin (as determined by the name passed in the
command) and flags (if applicable) are found.

BUG=b:196631044
TEST=tast run [user@]host[:port] firmware.FwupdPluginStartup

Change-Id: I99a07e7d9ac887d6dec607d50ec76600e7159ae8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3094786
Tested-by: Daniel Campello <campello@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Daniel Campello <campello@chromium.org>