tast-tests: Fix CUJ close window flake
ash.Window.CloseWindow trigger window close but does not
wait for the window close to finish. As a result, MeetCUJ
could try to change window state on a closed window and
flakily fail. This CL makes CloseWindow wait for window
closure to fix the flake in MeetCUJ and a similar case
in TaskSwitchCUJ.

BUG=chromium:1239283
TEST=tast run $DUT ui.MeetCUJ.lacros_4p

Change-Id: Id374b270ad5ff2805a97d020a8c12c40400e12b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3088356
Tested-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Reviewed-by: Renuga Nanjappan <rnanjappan@chromium.org>
Commit-Queue: Xiyuan Xia <xiyuan@chromium.org>