fwupd: restart fwupd upstart in inhibit suspend test
Background: updateChecker is a function that determines whether the
fake device in fwupd has been updated and a reinstall must be run,
or else to tell the user to update.

This removes updateChecker in favor of restarting the fwupd upstart
script, which will allow the fake device to be updated every time
without checking for whether or not it is possible.

BUG=b:196402627
TEST=tast run [user]@host:[port] firmware.FwupdInhibitSuspend

Change-Id: I95b1bed028247dd72940046f1b4c0ac4396c4fa5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3091625
Tested-by: Twain Byrnes <binarynewts@google.com>
Reviewed-by: Benson Leung <bleung@google.com>
Reviewed-by: Daniel Campello <campello@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Daniel Campello <campello@chromium.org>