Graphics.IGT: Fix kms_flip and kms_vblank
1. Extend the timeout of kms_vblank for 15 mins.

2.Few IGT tests like kms_flip run tests in combination with suspend.
For this, waking up is achieved using the RTC wake-up alarm.

If any other service is holding /dev/rtc, the IGT test fails to take the lock and set a wake up alarm. Hence, it is required to stop the service which is currently holding the RTC lock before running the IGT test.
In case of chrome, tlsdated holds rtc lock. This patch stops the
service before IGT tests are started and stop it when it's done.

BUG=b:193034528
TEST=cros_sdk tast run VOLTEER graphics.IGT.kms_{flip,vblank}

Change-Id: I5428d5aea1a091c3f4a94ffc6ea8ef6a2a0bf622