tast-tests: Stabilize filemanager.LaunchFromShelfContextMenu
The test was incorrectly polling for a single Files app window without
checking first that one was open. The Files app window doesn't open in
time for the poll to first run and it fails as no window is open. This
adds a check to make sure the window is open then ensures that the
window is the only one open (and stays open) for 5 seconds.

BUG=chromium:1238131
TEST=tast run <DUT-IP> filemanager.LaunchFromShelfContextMenu

Change-Id: I16c9d328669321f054fcaec9e7f50a48da73d64e