wifi: ChannelScanDwellTime: Ignore beacons sent before probe request
When validating scan behavior the ChannelScanDwellTime will compare all
beacons transmitted by the test AP. Because scan results are not
available during the period the probe request is sent, some initial
beacons may be considered missed if the test does not account for this
behavior:

	Found 82 test beacons between "wifi_CSDT_z0zlz_000001A4" and "wifi_CSDT_z0zlz_000001F9"
	Error at channel_scan_dwell_time.go:223: Testcase #1 (ch9): Missed 4 beacons: [wifi_CSDT_z0zlz_000001A4 wifi_CSDT_z0zlz_000001A9 wifi_CSDT_z0zlz_000001AA wifi_CSDT_z0zlz_000001AB

Instead, consider the dwell time as the time spent in channel after the
probe request. This will avoid flake related to beacon discontinuity at
the start of a scan before the probe request, and will more accurately
measure the dwell time.

BUG=b:192587894
TEST=`tast run -var=router=$ROUTER_IP -var=pcap=$PCAP_IP $DUT_IP wifi.ChannelScanDwellTime` succeeds

Change-Id: Ib4d4f4837557e505b83ba8a893d516b45cee2438