video: skip MemCheck on rockchip devices (scarlet, kevin etc)
Rockchip devices, despite having a 4.4 kernel and supporting
NV12Overlays, don't have a DRI debugfs output for DRM framebuffers,
so we don't have a way to "count" video buffers in this way.
This CL skips those tests on those platforms.

This CL also removes the superfluous per-test ExtraHardwareDeps
(since it's already in HardwareDeps).

BUG=b:195755581
TEST=tast run <kevin IP> video.MemCheck.h264_hw is skipped

Change-Id: I1495f89465e97ec241ba946d2c4d91526523f225