ChromeOS Real-time Reporting: Add Integration Test
This CL adds a Tast integration test that covers the interaction between the
ChromeOS client and the Reporting API server.

The test logs ChromeOS in as an enterprise user enrolled in an org that
enforces Cloud Reporting. At the same time, the test also configures the
ChromeOS client to report a heartbeat event. Finally, upon signing in
and starting Chrome, the test calls the Reporting API Server's
ChromeReportingDebugService.LookupEvents http endpoint to verify that
the server successfully received the heartbeart event.

BUG=chromium:1194164
TEST=ran test on a physical device.
Change-Id: I90f5425bb14565f5c5484329a0d7519b2e7ee66d