tast-tests: Remove a fixture for appsplatform.WebAPK
As arcBootedWithWebAppSharing is only used by appsplatform.WebAPK and
its only purpose is to pass an additional feature flag, let's just start
up Chrome and ARC directly from the test without a fixture.

Also, arcBooted will be updated to launch UI Automator by default, so
it is better not to use a fixture for appsplatform.WebAPK, because it
want to launch and stop UI Automator on its own. (See the comment in
clickShareTextButton()). arcBooted CL: https://crrev.com/c/3005502

TEST=tast run hatch-r-DUT appsplatform.WebAPK.vm
BUG=b:190014347

Change-Id: I67e2134332f4286097888c0d03f75568db8208b8