tast: Improve UIInput.gain to test with flexible low/high gain
Quicksettings only support coarse mic gain slider value adjustment (+-10
per step). For some reasons the initial slider value may not be 50,
which makes it infeasible to adjust the value to 30 and fail the test.

This CL improved the test by allowing the testing slider value within a
flexible range [30,40). And calculate the expected gain with regard to
the low/high value it adjusted.

BUG=b:196775829
TEST=tast run $ELM_IP audio.UIInput.gain

Change-Id: I5c726d46ab03c06b090d581abbaf8cd83d72abfa