tast: Increase black detection threshold for drm playback
During HW DRM playback, we do a screenshot and check to see if it is all
black to further verify that HW DRM was actually used. We then see if
enough of that covers the screen. Testing has shown us this threshold
should be lower (results have come in at 0.94), and it is being changed
to 0.92.

BUG=b:196732697
TEST=tast run DUT video.PlayDRM.*

Change-Id: I8d0d0f9f7e588bf8eae541cf2221e78cd0355c6c