tast-tests: Verify Files app is pinned by default before launching
The Files app icon is not default pinned to the shelf on kukui devices.
This test is verifying a regression in the New window shelf launch path
so this uses autotestPrivate.getDefaultPinnedApps API to verify that the
Files app should be pinned by default on this device before trying to
launch it.

BUG=chromium:1238131
TEST=tast run <kukui-device> filemanager.LaunchFromShelfContextMenu

Change-Id: Id4302923a5dc8baae54640653c1fde84377c0d7a