crash: Improve ECCrash test
Use the firmware precondition NormalMode, because it will try to make
sure the DUT is booted after the test. This requires a symlink be
created to access the fw-testing-configs.

Close the gRPC client before crashing the machine, as it can be very
slow to close if the connection is already dead.

Remove the CCD watchdog before crashing. This is likely the source of
instability in the test. Any reboot of the EC can cause servod to exit
or silently become read only if the CCD watchdog notices the EC is down.

BUG=b:187792616
TEST=tast run $DUT crash.ECCrash # on leona

Change-Id: I8a35d43960835393d4995d7125017bd24586e7de
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3101452
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>