tast-tests: Update devcore tests for compressed devcore files
In CL:3101776 we've started gziping devcoredump files because they're
sometimes too big to be uploaded to the crash/ server. Update tests for
the new file extension. Also update the intel fw dump test for the new
file size, which will be less than 1MB now (and looks to be about 800kB
in practice).

BUG=b:172987695
TEST=tast run $DUT crash.*

Cq-Depend: chromium:3101776
Change-Id: If92a2c7f251172fa26a1198cc29b7725f9f9e2b8