tast-tests: stabilize handwriting test
A previous CL used short timeout to limit the time for warmup, it
introduces flakiness in the cleanup of releasing mouse, which may
happen when drawing strokes.

This issue doesn't happen when a warmup file is used for warming up.

Revise the warmup process to use the handwriting file but only draw
the first stroke.

Refactor the drawStrokesFromFile func to be able to draw a specific
number of strokes.

Add DrawFirstStrokeFromFile to draw the first stroke.

BUG=b:187795148
TEST=tast run <DUT> inputs.VirtualKeyboardHan*

Change-Id: Ie5be9312be5b97be5244d4d3375a7dbcb3ec4ece
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3105686
Tested-by: Alvin Jia <alvinjia@google.com>
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Commit-Queue: Alvin Jia <alvinjia@google.com>