tast-tests: Parameterize the TabletOperations test with touch size.
Add a parameterized version of the test to run on model "kohaku"
specifically, and add logic in TabletOperations to take that parameter and parse it.

Add a method to the single touch event writer to update the internal
touchMajor/touchMinor representation, which updates simulated data on
device.

BUG=b:196859354
TEST=tast run $KOHAKU ui.TabletOperations is skipped. tast run $KOHAKU ui.TabletOperations.kahuku executes and passes.

Change-Id: Ifef9cae6a34ca388ea531d131772c8683170a965
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3105693
Commit-Queue: Rob Schonberger <robsc@chromium.org>
Tested-by: Rob Schonberger <robsc@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>