CheckIntelSARTable: Remove badModels functionality
Previously, CheckIntelSARTable had a mechanism to skip the test on
devices that had known-bad SAR tables. All known-bad models have now
been updated and fixed, so the badModels list is empty. Going forward, we
won't use this mechanism for devices with bad SAR tables, so remove the
functionality from the test.

BUG=b:181049667
TEST=tast run $DUT wifi.CheckIntelSARTable

Change-Id: Ib284efb2f729cd8e7e92d561f23c180309ec74f5