tast-tests split GetBluetoothPowered into GetBluetoothBootPref
The wifi.PersistenceBluetoothSansWifi as well as
wifi.PersistenceWifiSansBluetooth were failing on the function
GetBluetoothPowered due to timeout on the context.

It's unclear whether the Persistence tests are failing on obtaining the
boot preference from the DUT or from getting the expected adapter
status from the device. We split the function to gain visibility into
these functions as well as checkpoint the polling so that if the issue
was on waiting for the correct adapter status, we wouldn't have to keep
spawning new chrome sessions (which is done when retrieving the boot
pref)

BUG=b:190755771
TEST=tast run ${DUT} wifi.PersistenceBluetoothSansWifi and
wifi.PersistenceWifiSansBluetooth

Change-Id: If89be75dab82111845136a8c49ba458b50440f5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3108551
Reviewed-by: Kevin Lund <kglund@google.com>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Commit-Queue: Billy Zhao <billyzhao@chromium.org>
Tested-by: Billy Zhao <billyzhao@chromium.org>