faft: Speed up boot mode switcher
The 5 second delay in SetUSBMuxState is only triggered by the boot mode
switcher, which doesn't need it because it immediately follows the call
by waiting for the machine to boot.

When booting from rec->dev, try the menu keys. That avoids a dev screen
timeout in most cases, and shouldn't cause any harm if rebooting in
recovery ends up in normal mode.

Log more keypresses and sleeps.

In the keyboard dev switcher, keep pressing CTRL-D until the machine
boots. The sequence is Ctrl-D, ENTER, wait a long time for the power
wash, Ctrl-D, and if you don't press that last Ctrl-D there is a 30s
timeout before it finishes booting.

BUG=b:174846911
TEST=tast run ${DUT_HOSTNAME?} firmware.Fixture* # eldrid

Change-Id: I6d5420fb54fc154fd00af0bd365e482ac206af35
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3114983
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>
Reviewed-by: Renuga Nanjappan <rnanjappan@chromium.org>