typec: Add timing delays to improve Basic test robustness
Some of the servo commands require delays in order to allow the
resulting PD communication and voltage changes to settle. Not doing so
leads to flaky behaviour on certain TCPCs which lead to test failure.

Insert the appropriate delays to improve test robustness.

Also, remove an extra ccd_keepalive_en write, which wasn't necessary.

BUG=b:197121786
TEST=typec.Basic passed 50/50 times on delbin & eldrid.

Change-Id: I049a7fdce9b67401d1204f9b0d0ebd48df808411