tast-tests: Use real files in ArcChromeWebApkTest
Previously we were using pipes to implement a ContentProvider purely
in-memory. This appears to have been causing errors when reading the
files on rvc-arc.

Instead, this switches the ContentProvider to write the files to disk
and then return a file descriptor for the real file, which is a more
realistic way to implement the behavior.

BUG=chromium:1226730
TEST=tast run $DUT appsplatform.WebAPK.vm

Change-Id: Ib03fd1fc2adc17da2ef4ef838e3cddb4c38d9d60
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3115735
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>
Commit-Queue: Tim Sergeant <tsergeant@chromium.org>
Tested-by: Tim Sergeant <tsergeant@chromium.org>