arc: Reuse keyboard in arc.ResizeLock
For devices that don't have a keyboard attached, a virtual keyboard
is created in the test, but as it's a bit costly operation and the
test insert key events many times, this may lead to a timeout of
the test.

With this CL, the test only creates a keyboard instance once at the
beginning of the test and reuses it throughout the execution.

BUG=b:194156232
BUG=b:194321671
TEST=arc.ResizeLock

Change-Id: Iae20c9460f76f22669bd69c023c1db48a50782f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3115692
Reviewed-by: Toshiki Kikuchi <toshikikikuchi@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Kazuki Takise <takise@chromium.org>
Auto-Submit: Kazuki Takise <takise@chromium.org>
Commit-Queue: Kazuki Takise <takise@chromium.org>