nearbyshare: add support for controlling the native sharing UI
Currently we navigate to chrome://nearby/share in a browser to start the
sending flow. This is sufficient for testing the basic sharing flow, but
ARC++ sharesheet testing requires using the actual UI, since it will be
launched from the ARC++ entry point. The nearbyshare library can support
this by using cr.NewConnForTarget to connect to the underlying Chrome
associated with the UI.

BUG=b:197561642
TEST=tast run <DUT> nearbyshare.CrosToPhoneHighVis*11kb, with local
modification to start the sharing UI from the native sharesheet.

Change-Id: Ie18fde12c03fc8fa76c472fe87787050ed9f6e4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3116491
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Alan Ding <alanding@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>