typec: Add CC offs before toggling servo dts
Toggling the servod DTS state involves an implict cc-off and cc-on in
quick succession. On certain hardware, this wedges the Cr50 enumeration
of the DUT, which leads to servod crashing, thus causing the test to
fail.

So, insert a CC-off followed by a Vbus discharge delay before touching
src dts. This has the benefit of not making the CC-off --> CC-on
transition so sudden.

BUG=b:189595601
TEST=typec.Basic passes 100/100 times on elemi and lindar.

Change-Id: I9e59bfc240c86058434c56c23b5d17ebd3ef0425