camera: Collect JS error when fails to wait for video active
It's a quite common error when tests hit the timeout when waiting
for video active. However, it cannot call app.Close() successfully now
since the context is already timeout.

This CL adds a cleanup context for closing the app if it fails, so that
we can collect the JS errors properly if those are the root cause.

BUG=b:197750983
TEST=tast run [DUT] camera.CCAUISmoke.real

Change-Id: I8ef5561bac95d661979ba8aa34f1ad3b3202e82e