arcappcompat: Add standardized mouse right click test
Adds a button that handles right clicks to the Mouse application and
implements the tast test to perform the test. Additionally, abstracted
out logic in the mouse file to allow additional buttons to be registered
and cleaned up code in Android application to conform to AOSP standards.

BUG=b:194900373
TEST=push apk && tast run ${HOST} arc.StandardizedMouse*

Change-Id: Ic231b67a925d27d5d1d08acecf18ac12b6a036b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3118389
Tested-by: David Welling <davidwelling@google.com>
Commit-Queue: David Welling <davidwelling@google.com>
Reviewed-by: Tetsui Ohkubo <tetsui@chromium.org>
Reviewed-by: Mohamed Omar <mohamedaomar@google.com>