scanapp: Handle printer notifications
When Chrome is done setting up the virtual USB printer, it displays a
notification.  If this notification comes at the exact wrong time, it
can cover the Scan button and steal the intended mouse click.  To avoid
this, wait for the notification to come up before launching the app and
then close all the notifications before attempting to scan.

BUG=b:188591629
TEST=50 iterations of tast run scanapp.* on octopus

Change-Id: I5c7f8db5c347594aadd560d9c3a7296e748e740c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3119968
Reviewed-by: Kyra Moed <kmoed@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>