arcappcompat: Update Tast test for Docusign
Docusign is failing on launch. It is observed that
with the current version of Docusign, the app is
failing on launch on some models due to different
timeouts needed for different models to get into
the enter email address page. With the older version
of app, email address page was on web view had to use
the keycode to enter the email address.  It is observed
that on Bluebird, Berknip, tab key and email address
are entered already before it reaches the email address field page.
In the current version of Docusign, email address page id's
are available. Updated the code with a new IDs, which will fix
the launch failure.

BUG=b:193840357
TEST=tast run <eve> arcappcompat.Docusign

Change-Id: I1972936675e0625cbaf96ba5b7707c58a1dc9c04