arcappcompat: Add standardized touchscreen zoom test
Update the Android application to handle zooming in, and zooming out on
an element and add a corresponding test to handle the zoom functionality
on the touchscreen. Additionally, added helper methods in the touch
input file to support pathing between two points, and moving multiple
touches at the same time.

BUG=b:194900373
TEST=push apk && tast run ${HOST} arc.StandardizedTouch*

Change-Id: I2491fe78419ed1a18f832aacd3a0332d5ebccb37