arc: Wait for content to be drawn before checking black flash
arc.BlackFlash checks if there's any black part in app content.
In order to check the app content, the test has the logic to wait
for the content to be drawn (to be blue), but as it's checked
after the actual black flash check, the test ends up checking the
screenshot without waiting the app content.

BUG=b:194156232
TEST=arc.BlackFlash.vm

Change-Id: Ice1fc7a1dec05d018084cfae65e1fe609744d010
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3121691
Tested-by: Kazuki Takise <takise@chromium.org>
Auto-Submit: Kazuki Takise <takise@chromium.org>
Commit-Queue: Kazuki Takise <takise@chromium.org>
Reviewed-by: Alvin Jia <alvinjia@google.com>
Reviewed-by: Oleh Lamzin <lamzin@google.com>