hwsec: test login method with LECredential in LockToSingleUserMountUntilReboot
Currently the test does not exercise the login method with LECredential.
Add one more key that uses LECredential (PIN) for both users, and verify
the same test logics work on LECredential too. On devices that do not
support PinWeaer, only test the first and second key (which is same as
the current test).

Also, the constants FirstPin/SecondPin/PinLabel are currently used as a
second set of password instead of PINS. Add constants
FirstPassword2/SecondPassword2/Password2Label to replace the original usages,
and use the pin-related constants as PINs. Fix related comments/messages.

BUG=b:192822451
TEST=tast run $DUT hwsec.*

Change-Id: I4dfb5ecf61f96ac93f04a7ba287b2d1b91f771ed