tast-test: Add Large Paper Scan tast test
Implements a tast test for the update to the scanapp allowing larger
paper sizes to be scanned.
Due to the overlapping nature of this new test with the scanapp.Scan
test, creates a single function to run the bulk of the test. Each
individual test sets up its parameters for running, then calls the
shared function.

BUG=b:192658314
TEST=scanapp.LargePaperScans

Cq-Depend: chromium:3123005,chromium:3129184
Disallow-Recycled-Builds: test-failures
Change-Id: Ic61a135585152c813e430b38cc10bc0bab30dd57