wifi: RRMBeaconReport: wait for scan when applicable
For beacon report requests, active and passive and passive modes will
cause the DUT to perform an active or passive scan respectively. Use
WpaMonitor to wait for the scan result event so that we don't
prematurely kill the packet capture and miss the report.

Note: ideally we'd wait for the beacon report event on the hostapd side,
but we currently don't have a wait to do that, and this should be good
enough.

BUG=b:197914719
TEST=ran test several times, doesn't seem to flake anymore

Change-Id: Ic001672a089faef21abeb70bb03ee308edaa4ced
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3124275
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Frank Gorgenyi <frankgor@google.com>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>