faft: New pattern and longer timeout
Treat x86 machines as AP booted when 'Executing host reboot command'
is seen, that indicates a host command sent from the AP just like
'HC 0x' does.

Allow x86 machines to boot in 1.5s. That should let all the machines in
the lab pass.

Move this test to unstable as well, so people won't run it.

BUG=b:197776884, b:172227463
TEST=on eldrid, voema, eve

Change-Id: Ie86876baad45f8c5439af65097abf37b5f9446b6