Don't fail CB->CB tests when a log file is missing.
Currently the tests hang if a log file doesn't exist. This CL fixes that
hang and will allow the test to continue without failing if a log file
is not available.

There is currently an issue where /var/log/messages is not present and
instead a .tgz file is instead. Until that is fixed I will let the test
just log this condition and continue to pass if everything else is good.

BUG=b:197686791
BUG=b:198200585
TEST=nearbyshare.CrosToCrosHighVis.dataoffline_allcontacts_jpg11kb with
file present and file not present

Change-Id: I73949be8b63ae652486cd04e7fee641633866eff