network.*: distribute Contacts:briannorris@
I'm not primarily working on shill/networking.

network.Ethernet8021X.*: this tests shill + wpa_supplicant integration
for Ethernet. It is most likely to hit the WiFi domain, I think, as the
WiFi team owns and modifies wpa_supplicant the most. Add the usual
champs list, with cros-networking@ as a secondary.

network.SupportedProtocols: this isn't really shill-specific; we just
want to make sure protocols are there. I can stick on the Contacts list,
as this is dead-simple, but put cros-networking@ as well.

stevenjb@ volunteered to pick up non-technology-specific shill tests,
which I believe the rest qualify (they're mostly exercising profile
behavior).

BUG=none
TEST=none

Change-Id: I5e17b0bf2a3b0d95be55a542a52dd9085fe21201