tast-tests: Relax arc.WindowState.*_stress*
This CL bumps up the test timeout for arc.WindowState.*_stress* from 4 min to 10 min.
The original 4 min timeout is easily exceeded on slower devices (e.g. dedede/puff) and the running time is around 8 min on drawlat.
This CL also reduces the iteration count from 25 to 20 to shorten the running time.

BUG=b:179525662
TEST=arc.WindowState.*_stress*

Change-Id: Ibbdda489bcbed5104d83ce6e2aa801ddb748fb9d