platform_decoding: Remove multi-line error message
Today when a video.PlatformDecoding test case fails to decode a file,
the entire test binary output is logged in the failure message. This is
not best practice for tests.

This change moves the logging of the output of a failing test case from
the failure message to the test context log.

BUG=none
TEST=tast run video.PlatformDecoding.v4l2_vp9_* on Trogdor (has errors)

Change-Id: Ic65c30468eeaaa3a2c8b6d317cedd6401840b03d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3135096
Tested-by: Nathan Hebert <nhebert@chromium.org>
Commit-Queue: Nathan Hebert <nhebert@chromium.org>
Reviewed-by: Patrick Meiring <meiring@google.com>