tast-tests: stabilize VK Float tests
The test is still flaky on some boards, and the reason has been
identified: it checks whether the access points panel is shown before
the keyboard is rendered.

Revise it to wait for keyboard to render before checking the
appearance of the access points panel.

Verified on a cyan device with 50 passes in a row.

BUG=b:196880784
TEST=tast run <DUT> inputs.VirtualKeyboardFloat*

Change-Id: I7992b0906cb61de7fed51e312bf54d36edad237e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3132174
Tested-by: Alvin Jia <alvinjia@google.com>
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Commit-Queue: Alvin Jia <alvinjia@google.com>