camera: Send intent only when clicking button in CCAUIIntent
Originally the intent is sent when the test app launching. On ARCVM
under tablet mode, there is an issue that the window might show up after
onCreate()/onResume() is triggered, which make the launched CCA lose the
focus and cause the test fail.

As a workaround, we add a button in the test app, and only send the
intent when the button is clicked so that we can make sure test app
window shows first.

BUG=b:183691889
TEST=tast run [DUT] camera.CCAUIIntent.vm (on Kohaku)

Change-Id: I62a2006121d785af8b565c733d223a806ee81231
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3131254
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Reviewed-by: Shik Chen <shik@chromium.org>
Commit-Queue: Wei Lee <wtlee@chromium.org>
Tested-by: Wei Lee <wtlee@chromium.org>
Auto-Submit: Wei Lee <wtlee@chromium.org>