tast-tests: Add arc.MediaScanPerf test
Measure the elapsed time while media scan to sdcard / MyFiles directory
in parametrized test.

These directories are mounted and scanned just after ARC is started. Therefore,
in order to run media scan again, we test the performance following the steps below.

1. create a lot of image files under it
2. wait for the newly created files to be added to MediaStore database
3. delete them from the database
4. unmount the target directory
5. mount it again

Step 1 is needed to clarify the the difference in elapsed time due to
poor performance. Also, inotify in ArcFSWatcherService watches filepath
change and issues a request about asking MediaStore to scan the files.
Therefore, if you don't take step 2~3, the newly created files will be
in the status of scanned when the directory is remounted and full-volume
media scanner is triggered. To avoid such an obstruction, we need to
delete the created files from MediaStore database and let MediaScanner
scan whole the target directory.

BUG=b:198243699
TEST=$ emerge-${BOARD} tast-local-apks-cros
TEST=$ cros deploy --root=/usr/local --board=${BOARD} <DUT> \
TEST=  tast-local-apks-cros
TEST=$ tast run <DUT> arc.MediaScanPerf*

Change-Id: Iec604f07204089cd1983f0c2f3dba3cb21f2919f