tast-tests: Add arc.MediaScanTime test
Measure the elapsed time while media scan to sdcard / MyFiles directory
in parametrized test.

These directories are mounted and scanned just after ARC is started. Therefore,
in order to run media scan again, we test the performance following the steps below.

1. unmount the target directory
2. create a lot of image files under it
3. mount it again

step 2 is needed to clarify the the difference in elapsed time due to
poor performance.

BUG=b:198243699
TEST=$ emerge-${BOARD} tast-local-apks-cros
TEST=$ cros deploy --root=/usr/local --board=${BOARD} <DUT> \
TEST=  tast-local-apks-cros
TEST=$ tast run <DUT> arc.MediaScanTime*

Change-Id: Iec604f07204089cd1983f0c2f3dba3cb21f2919f