memory: Ignore unused zones
Kernel commit 1a8e1cef7603 ("arm64: use both ZONE_DMA and ZONE_DMA32")
introduced ZONE_DMA which contains the lower 1G of memory, so it's possible
that ZONE DMA32 is not used.

Add a function to check if DMA32 is not used and don't report error in
this case.

BUG=b:194506209
TEST=pass tast.multivm.Lifecycle.host on kukui-kernelnext

Change-Id: If41a1a160284d2880d5e8bf24bfecd28042d7e46
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3135112
Tested-by: Hsin-Yi Wang <hsinyi@chromium.org>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Reviewed-by: Charles William Dick <cwd@google.com>
Commit-Queue: Hsin-Yi Wang <hsinyi@chromium.org>