tast-tests: Add toolkit and kiosk tests
Add 'Toolkit' and 'Kiosk' tests along with a Fixture 'ensureToolkit'.

The Fixture ensureToolkit ensures the toolkit is installed with a reboot. The
Toolkit tests whether the toolkit is available in system and the Kiosk tests
whether browser is running with the debug port 9222 enabled, then checks is it
displaying or loading a test page.

BUG=b:191446203
TEST=tast run <DUT> factory.Toolkit factory.Kiosk

Change-Id: I15c8366cab3fe91c8f1c602392df2afc514f5562