tast-tests: Add toolkit and kiosk tests
Add 'Toolkit' and 'Kiosk' tests along with a Fixture 'ensureToolkit'.

The Fixture ensureToolkit ensures the toolkit is installed with a reboot, and
performs uninstallation of toolkit with a reboot at the end of all tests
require this fixture. The Toolkit tests whether the toolkit is available in
system and the Kiosk tests whether browser is running with the debug port 9222
enabled, then checks is it displaying factory test page.

BUG=b:191446203
TEST=tast run <DUT> factory.Toolkit factory.Kiosk

Change-Id: I15c8366cab3fe91c8f1c602392df2afc514f5562
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3139599
Tested-by: Sung-chi, Li <lschyi@google.com>
Commit-Queue: Sung-chi, Li <lschyi@google.com>
Reviewed-by: Meng-Huan Yu <menghuan@chromium.org>
Reviewed-by: Matthew Warton <mwarton@google.com>