(BeTo) Refactor Mobly snippet client into its own package
I implemented a Mobly snippet client as part of the Nearby Share
automation effort, since we use a snippet to control Nearby Share on
Android. We need snippets for testing other Better Together features
such as Phone Hub and Smart Lock, so this CL extracts the Mobly snippet
code from a Nearby-specific package to its own package so we can use it
for other feature automation.

BUG=b:192473324
TEST=tast run <dut> nearbyshare.*Phone*

Change-Id: I42979f0d60952c750ae1d8de3311feff924f643d