crostini: Move ui.gaiaPoolDefault back to Vars for most tests
In CL:2919549 this variable was made a hard dependency for all tests
which used it. However, most crostini tests do not require this
variable and only declared a use of it as part of a partially
completed migration to making all tests use real gaia accounts where
available.

This has caused all crostini tests to stop running on the chrome bots,
when really only three should actually be blocked. This CL corrects
this for all tests that do not use the StartedByDlc*Gaia preconditions
(which do require this variable to be available).

BUG=chromium:1245932
TEST=CQ

Change-Id: I8d91428bb06eb7efe51f6a8e776aef894a1d7514
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3139971
Reviewed-by: Judy Wu <jinrongwu@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
Tested-by: Fergus Dall <sidereal@google.com>