video cuj: ensure YouTube app not opened when testing YouTube web
It has been seen that Chromebook sometimes try to open YouTube links in
YouTube app instead of staying in the Chrome browser. This has caused
YouTube app to be opened unintentionally when testing YouTube web. This
commit tries to resolve this issue by instructing DUT to stay with
YouTube web when prompted to switch to YouTube app.

BUG=b:199165566
TEST=tast run <vars> <dut> ui.VideoCUJ2.*

Change-Id: I299b0aec0d1e6f6b3082a9365c82fc9999421231
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3144977
Reviewed-by: Judy Wu <jinrongwu@google.com>
Tested-by: Xibin Liu <xibin@google.com>
Commit-Queue: Xibin Liu <xibin@google.com>